requirements/base.txt
blinker==1.4
cffi==1.5.2
cryptography==1.2.3
Flask==0.10.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.1
Flask-Uploads==0.2.0
Flask-WTF==0.12
idna==2.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pyasn1==0.1.9
pycparser==2.14
PyJWT==1.4.0
six==1.10.0
SQLAlchemy==1.0.12
Werkzeug==0.11.4
WTForms==2.1
